Update GPL -> MIT
After a discussion with various folks about licensing, I think it's the case that even though we use circom code, we don't fork its compiler, and directly import the GPL licensed circomlib library (with their GPL license), and don't fork that code. So it seems to me that our whole repo can safely be MIT licensed.